#40396 64bit環境でMIDI入力を受け付けない問題を修正。callback内のint→IntPtr置き換えが漏れていた。

+++ FDK/コード/00.共通/CWin32.cs
@@ -808,6 +808,6 @@ namespace FDK
 
 		// Win32 メッセージ処理デリゲート
 
-		public delegate void MidiInProc( IntPtr hMidiIn, uint wMsg, int dwInstance, int dwParam1, int dwParam2 );
+		public delegate void MidiInProc( IntPtr hMidiIn, uint wMsg, IntPtr dwInstance, IntPtr dwParam1, IntPtr dwParam2 );
 	}
 }
